removed charitywizard (incorrect name), copied relevant code to the working activity
