+++ DBAndroid/app/src/main/AndroidManifest.xml
@@ -14,12 +14,11 @@
         android:usesCleartextTraffic="true">
 
 
-        <activity android:name=".charity.CharityWizad"></activity>
+
         <activity android:name=".donor.DonorWizard" />
         <activity android:name=".common.ApplicationActivity"></activity>
         <activity android:name=".common.ListingActivity" />
         <activity android:name=".charity.CharityWizard"></activity>
-        <activity android:name=".donor.donorWizard" />
 
         <activity android:name=".RegistrationResultActivity" />
         <activity android:name=".util.simpleResult.SimpleResultActivity" />
