+++ DBAndroid/app/src/main/res/layout/activity_charity_wizard.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context=".charity.CharityWizad">
+    tools:context=".charity.CharityWizard">
 
     <Button
         android:id="@+id/postButton"
@@ -116,13 +116,13 @@
         app:layout_constraintBottom_toTopOf="@+id/charwizLocation"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
-    android:layout_marginStart="16dp"
-    android:layout_marginTop="8dp"
-    android:text="@string/location"
-    android:textAllCaps="true"
-    android:textStyle="bold"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/charwizDateTime" />
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/location"
+        android:textAllCaps="true"
+        android:textStyle="bold"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/charwizDateTime" />
 
     <TextView
         android:id="@+id/eventDateTime"
@@ -242,7 +242,8 @@
         android:layout_width="388dp"
         android:layout_height="34dp"
         tools:layout_editor_absoluteX="7dp"
-        tools:layout_editor_absoluteY="412dp">
+        tools:layout_editor_absoluteY="412dp"
+        tools:ignore="MissingConstraints">
 
         <android.support.design.widget.TextInputEditText
             android:layout_width="match_parent"
